Refactor(Horarios): Passado as funcoes para um arquivo separado
Ajustado o alinhamento com os demais formulários da aplicação
